API is now Object Oriented
Finally updating the API.
This has been a long time coming because I've abandoned the original as it was just to messy and replaced it with a more OOP approach.
Currently the API only works with one dataset (as I still haven't got more from Binoy) and it can handle Regression of Viscosity to a dataset and print the results.
Now that the code base is much much more managable, future updates should come quickly (hopefully).

Plan is also making the Parser/Director Object Oriented, so working with it is easier.